2025.11.26
Even with the  "cache busting" method, cloudflare cache still resulted in too many issues where trending collections were incomplete or out of order. UMTK now outputs the IDs to ensure the collection in Plex is complete. To be able to sort a trending collection, the `add_rank_to_sort_title` variable was added to prefix the rank to the sort_title.